+++ Assets/Scripts/Enemy/Enemy_GiantBee.cs
@@ -33,7 +33,7 @@ public class BeeEnemy : MonoBehaviour
     transform.position=Vector2.MoveTowards(transform.position, player.transform.position, speed*Time.deltaTime);
   }
 
-  void OnHealthChanged(float current, float max)
+  void OnHealthChanged(int current, int max)
   {
     Debug.Log($"Health changed {current} {max}");
     if (current <= 0)
